rename component & generalise view/direction
